modularise rag code more
