+++ data-pipeline/src/pipeline/rag/main.py
@@ -1,8 +1,5 @@
-import itertools
-import os
 import time
-from multiprocessing import Pool
-from typing import Dict, List, Optional, Tuple
+from typing import List, Optional
 
 import pandas as pd
 
@@ -11,145 +8,10 @@ from pipeline.utils.database import insert_metric_rag
 from pipeline.utils.log import setup_logger
 from pipeline.utils.storage import get_blob, write_blob
 
-from .calculations import RAG_RESULT_COLUMNS, CategoryColumnCache, process_single_urn
+from .loader import load_school_data_and_comparators
+from .engine import _run_rag_computation_engine
 
 logger = setup_logger(__name__)
-
-# --- Worker functions for multiprocessing ---
-
-_WORKER_GLOBALS = {}
-
-
-def init_worker(processed_data: pd.DataFrame, comparator_map: Dict[str, dict]):
-    """
-    Initializer for each worker process. Caches data once per process.
-    """
-    global _WORKER_GLOBALS
-    logger.info(f"Initializing worker process: {os.getpid()}")
-    _WORKER_GLOBALS["processed_data"] = processed_data
-    _WORKER_GLOBALS["column_cache"] = CategoryColumnCache(processed_data.columns)
-    _WORKER_GLOBALS["comparator_map"] = comparator_map
-
-
-def run_rag_for_urn_worker(school_urn: str) -> list:
-    """
-    The task for a single worker: wraps the core single-URN processing logic.
-    """
-    try:
-        processed_data = _WORKER_GLOBALS["processed_data"]
-        column_cache = _WORKER_GLOBALS["column_cache"]
-        comparator_map = _WORKER_GLOBALS["comparator_map"]
-
-        target_school = processed_data.loc[school_urn]
-        # Get comparators specific to this URN from the map
-        comparators = comparator_map.get(school_urn, {})
-
-        results_generator = process_single_urn(
-            school_urn, target_school, processed_data, comparators, column_cache
-        )
-        return list(results_generator)
-
-    except Exception as e:
-        logger.error(f"Error processing URN {school_urn} in worker {os.getpid()}: {e}")
-        return []
-
-
-def create_empty_rag_dataframe() -> pd.DataFrame:
-    """Create an empty DataFrame with standard RAG result structure."""
-    return pd.DataFrame(columns=RAG_RESULT_COLUMNS).set_index("URN")
-
-
-def _run_rag_computation_engine(
-    processed_data: pd.DataFrame,
-    comparator_map: Dict[str, dict],
-    target_urn: Optional[str] = None,
-) -> pd.DataFrame:
-    """
-    Core RAG computation engine. Orchestrates serial or parallel execution.
-    This function is pure: it takes data in and returns a DataFrame, with no side effects like I/O.
-    """
-    all_results = []
-    empty_rag_df = create_empty_rag_dataframe()
-    column_cache = CategoryColumnCache(processed_data.columns)
-
-    if target_urn:
-        # --- Unparallelised ---
-        if target_urn not in processed_data.index:
-            logger.warning(f"Target URN {target_urn} not found in data")
-            return empty_rag_df
-
-        logger.info(f"Processing single target URN {target_urn} serially.")
-        try:
-            target_school = processed_data.loc[target_urn]
-            comparators = comparator_map.get(target_urn, {})
-            rag_generator = process_single_urn(
-                target_urn, target_school, processed_data, comparators, column_cache
-            )
-            all_results = list(rag_generator)
-        except Exception as e:
-            logger.error(f"Failed to calculate RAG for target URN {target_urn}: {e}")
-            return empty_rag_df
-    else:
-        # --- Parallelised ---
-        logger.info("Processing all schools in parallel.")
-        schools_to_process = list(comparator_map.keys())
-
-        if not schools_to_process:
-            logger.warning("No schools to process.")
-            return empty_rag_df
-
-        try:
-            num_processes = os.cpu_count() or 2
-            logger.info(f"Utilizing {num_processes} worker processes.")
-            pool_init_args = (processed_data, comparator_map)
-
-            with Pool(
-                processes=num_processes,
-                initializer=init_worker,
-                initargs=pool_init_args,
-            ) as pool:
-                results_list_of_lists = pool.map(
-                    run_rag_for_urn_worker, schools_to_process
-                )
-
-            all_results = list(itertools.chain.from_iterable(results_list_of_lists))
-        except Exception as e:
-            logger.error("Parallel RAG calculation failed: {e}", exc_info=True)
-            return empty_rag_df
-
-    return pd.DataFrame(all_results).set_index("URN") if all_results else empty_rag_df
-
-
-def load_school_data_and_comparators(
-    run_type: str, run_id: str, school_type: str
-) -> Tuple[pd.DataFrame, pd.DataFrame]:
-    """Load school data and comparator mappings from storage."""
-    if school_type == "maintained_schools":
-        school_filename = "maintained_schools.parquet"
-        comparator_filename = "maintained_schools_comparators.parquet"
-    elif school_type == "academies":
-        school_filename = "academies.parquet"
-        comparator_filename = "academy_comparators.parquet"
-    else:
-        raise ValueError(f"Unknown school type: {school_type}")
-
-    try:
-        school_data = pd.read_parquet(
-            get_blob("comparator-sets", f"{run_type}/{run_id}/{school_filename}")
-        )
-        comparator_data = pd.read_parquet(
-            get_blob("comparator-sets", f"{run_type}/{run_id}/{comparator_filename}")
-        )
-        logger.debug(
-            f"Loaded {school_type}: data_shape={school_data.shape}, comparators_shape={comparator_data.shape}"
-        )
-        return school_data, comparator_data
-    except Exception as e:
-        logger.error(f"Failed to load data for {school_type}: {e}")
-        raise
-
-
-# --- Entrypoint Functions (Responsible for I/O and calling the engine) ---
 
 
 def compute_rag(run_type: str, run_id: str, target_urn: Optional[str] = None) -> float:
